perf: reduce nats events (#9712)
