+++ src/lib.rs
@@ -13,7 +13,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#![feature(btree_cursors)]
 #![feature(variant_count)]
 
 #[cfg(feature = "enterprise")]
