+++ src/service/compact/flatten.rs
@@ -35,12 +35,12 @@ use config::{
     },
 };
 use hashbrown::HashSet;
-use infra::{file_list as infra_file_list, storage};
+use infra::{cluster::get_node_from_consistent_hash, file_list as infra_file_list, storage};
 use once_cell::sync::Lazy;
 use parking_lot::RwLock;
 use tokio::sync::{Semaphore, mpsc};
 
-use crate::{common::infra::cluster::get_node_from_consistent_hash, service::db};
+use crate::service::db;
 
 static PROCESSING_FILES: Lazy<RwLock<HashSet<String>>> = Lazy::new(|| RwLock::new(HashSet::new()));
