+++ src/service/compact/merge.rs
@@ -42,6 +42,7 @@ use config::{
 use hashbrown::{HashMap, HashSet};
 use infra::{
     cache::file_data,
+    cluster::get_node_by_uuid,
     dist_lock, file_list as infra_file_list,
     runtime::DATAFUSION_RUNTIME,
     schema::{
@@ -59,14 +60,11 @@ use tokio::{
 };
 
 use super::worker::{MergeBatch, MergeSender};
-use crate::{
-    common::infra::cluster::get_node_by_uuid,
-    service::{
-        db, file_list,
-        schema::generate_schema_for_defined_schema_fields,
-        search::datafusion::exec::{self, MergeParquetResult, TableBuilder},
-        tantivy::create_tantivy_index,
-    },
+use crate::service::{
+    db, file_list,
+    schema::generate_schema_for_defined_schema_fields,
+    search::datafusion::exec::{self, MergeParquetResult, TableBuilder},
+    tantivy::create_tantivy_index,
 };
 
 /// Generate merging job by stream
@@ -821,6 +819,9 @@ pub async fn merge_files(
         .values()
         .flat_map(|s| s.fields().iter().map(|f| f.name().to_string()))
         .collect::<HashSet<_>>();
+    // Keep original schema for index generation (has all stream fields)
+    let stream_schema_for_index = Arc::new(latest_schema.clone());
+    // Create filtered schema for parquet merging (only fields in parquet files)
     let latest_schema = Arc::new(latest_schema.retain(all_fields));
     let mut latest_schema_fields = HashMap::with_capacity(latest_schema.fields().len());
     for field in latest_schema.fields() {
@@ -978,6 +979,7 @@ pub async fn merge_files(
                     &index_fields,
                     &retain_file_list,
                     &mut new_file_meta,
+                    stream_schema_for_index.clone(),
                     &buf,
                 )
                 .await?;
@@ -1018,6 +1020,7 @@ pub async fn merge_files(
                         &index_fields,
                         &retain_file_list,
                         &mut new_file_meta,
+                        stream_schema_for_index.clone(),
                         &buf,
                     )
                     .await?;
@@ -1048,15 +1051,16 @@ async fn generate_inverted_index(
     index_fields: &[String],
     retain_file_list: &[FileKey],
     new_file_meta: &mut FileMeta,
+    latest_schema: Arc<Schema>,
     buf: &Bytes,
 ) -> Result<(), anyhow::Error> {
-    let (schema, reader) = get_recordbatch_reader_from_bytes(buf).await?;
+    let (_parquet_schema, reader) = get_recordbatch_reader_from_bytes(buf).await?;
     let index_size = create_tantivy_index(
         "COMPACTOR",
         new_file_key,
         full_text_search_fields,
         index_fields,
-        schema,
+        latest_schema, // Use stream schema to include all configured fields
         reader,
     )
     .await
